📕 chore: add transaction mocks to home correctly
